#781 [RENAME] 포인트 로그 페이지 폴더를 MyPage 하위의 pages 폴더로 이동
