+++ src/pages/MyPage/index.js
@@ -1,3 +1,5 @@
+export { default as PointLogsPage } from './pages/PointLogsPage/PointLogsPage';
+
 export { default as ChangePasswordPage } from './ChangePasswordPage/ChangePasswordPage';
 export { default as DeleteAccountPage } from './DeleteAccountPage/DeleteAccountPage';
 export { default as EditInfoPage } from './EditInfoPage/EditInfoPage';
