+++ src/pages/MyPage/pages/PointLogsPage/index.js
